--- src/main/java/LinusException.java
@@ -1,5 +1,0 @@
-public class LinusException extends Exception {
-    public LinusException(String message) {
-        super(message);
-    }
-}
